+++ dummyfs/file.h
@@ -18,7 +18,7 @@
 int dummyfs_truncate(oid_t *oid, size_t size);
 
 
-int dummyfs_truncate_internal(dummyfs_object_t *o, unsigned int size);
+int dummyfs_truncate_internal(dummyfs_object_t *o, size_t size);
 
 
 int dummyfs_read(oid_t *oid, offs_t offs, char *buff, size_t len);
@@ -27,7 +27,7 @@ int dummyfs_read(oid_t *oid, offs_t offs, char *buff, size_t len);
 int dummyfs_write(oid_t *oid, offs_t offs, const char *buff, size_t len);
 
 
-int dummyfs_write_internal(dummyfs_object_t *o, offs_t offs, const char *buff, unsigned int len);
+int dummyfs_write_internal(dummyfs_object_t *o, offs_t offs, const char *buff, size_t len);
 
 
 #endif /* _DUMMYFS_FILE_H_ */
